Add functional options, e.g. separate proc group
Add a new constructor that accepts functional options.  This (1.) does
not break the API and (2.) provides the option to run xz in a separate
process group.

The latter is only implemented for Linux and MacOS.  It is trivially
possible to implement that for other UNIX-alikes as well, but I strongly
believe in dog-fooding and Linux+MacOS is what I have and for which I
use this library.